[Refactoring]Use unsigned for in8 and out8
Both of these functions take a port number, which is always unsigned.
Using signed values may cause bugs.

+++ src/interrupt/mod.rs
@@ -5,28 +5,28 @@ use crate::queue;
 
 extern crate lazy_static;
 
-const PIC0_ICW1: i32 = 0x0020;
-const PIC0_OCW2: i32 = 0x0020;
-const PIC0_IMR: i32 = 0x0021;
-const PIC0_ICW2: i32 = 0x0021;
-const PIC0_ICW3: i32 = 0x0021;
-const PIC0_ICW4: i32 = 0x0021;
-const PIC1_ICW1: i32 = 0x00a0;
-const PIC1_OCW2: i32 = 0x00a0;
-const PIC1_IMR: i32 = 0x00a1;
-const PIC1_ICW2: i32 = 0x00a1;
-const PIC1_ICW3: i32 = 0x00a1;
-const PIC1_ICW4: i32 = 0x00a1;
+const PIC0_ICW1: u32 = 0x0020;
+const PIC0_OCW2: u32 = 0x0020;
+const PIC0_IMR: u32 = 0x0021;
+const PIC0_ICW2: u32 = 0x0021;
+const PIC0_ICW3: u32 = 0x0021;
+const PIC0_ICW4: u32 = 0x0021;
+const PIC1_ICW1: u32 = 0x00a0;
+const PIC1_OCW2: u32 = 0x00a0;
+const PIC1_IMR: u32 = 0x00a1;
+const PIC1_ICW2: u32 = 0x00a1;
+const PIC1_ICW3: u32 = 0x00a1;
+const PIC1_ICW4: u32 = 0x00a1;
 
-const PORT_KEYDATA: i32 = 0x0060;
+const PORT_KEYDATA: u32 = 0x0060;
 
-const PORT_KEY_STATUS: i32 = 0x0064;
-const PORT_KEY_CMD: i32 = 0x0064;
-const KEY_STATUS_SEND_NOT_READY: i32 = 0x02;
-const KEY_CMD_WRITE_MODE: i32 = 0x60;
-const KEY_CMD_MODE: i32 = 0x47;
-const KEY_CMD_SEND_TO_MOUSE: i32 = 0xd4;
-const MOUSE_CMD_ENABLE: i32 = 0xf4;
+const PORT_KEY_STATUS: u32 = 0x0064;
+const PORT_KEY_CMD: u32 = 0x0064;
+const KEY_STATUS_SEND_NOT_READY: u32 = 0x02;
+const KEY_CMD_WRITE_MODE: u32 = 0x60;
+const KEY_CMD_MODE: u32 = 0x47;
+const KEY_CMD_SEND_TO_MOUSE: u32 = 0xd4;
+const MOUSE_CMD_ENABLE: u32 = 0xf4;
 
 lazy_static::lazy_static! {
     pub static ref KEY_QUEUE: spin::Mutex<queue::Queue> = spin::Mutex::new(queue::Queue::new());
@@ -73,11 +73,11 @@ pub fn init_keyboard() -> () {
 
 pub extern "C" fn interrupt_handler_21() -> () {
     asm::out8(PIC0_OCW2, 0x61);
-    KEY_QUEUE.lock().enqueue(asm::in8(PORT_KEYDATA));
+    KEY_QUEUE.lock().enqueue(asm::in8(PORT_KEYDATA) as i32);
 }
 
 pub extern "C" fn interrupt_handler_2c() -> () {
     asm::out8(PIC1_OCW2, 0x64);
     asm::out8(PIC0_OCW2, 0x62);
-    mouse::QUEUE.lock().enqueue(asm::in8(PORT_KEYDATA));
+    mouse::QUEUE.lock().enqueue(asm::in8(PORT_KEYDATA) as i32);
 }
